feat: implement share link model, repository, and core security configuration with API testing index page

+++ main/src/main/java/com/cloudcvhub/repo/ResumeRepo.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository
 
-public interface ResumeRepo extends JpaRepository<Resume,Integer> {
+public interface ResumeRepo extends JpaRepository<Resume,Long> {
     List<Resume> findByUserIdAndIsDeletedFalse(Long userId); // lay danh sach CV cua user chua bi xoa mem
     Optional<Resume> findByIdAndUserIdAndIsDeletedFalse(Long id, Long userId);
 }
